+++ AppLovin/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+   * 6.5.4.0
+     * This version of the adapters has been certified with AppLovin SDK 6.5.4.
+     * Fix compiler warnings.
+     * Remove `userDeclinedToViewAd:` rewarded video delegate method as prompts have been removed.
+
    * 6.4.2.0
      * This version of the adapters has been certified with AppLovin SDK 6.4.2.